use dialog service to confirm deletion of diagnose item, refactor code to use data annotations from MVVM Toolkit

+++ App/ViewModels/DiagnoseViewModel.cs
@@ -1,18 +1,15 @@
 ﻿using System.Collections.ObjectModel;
-using System.ComponentModel.DataAnnotations;
 using System.Windows.Input;
 using App.Contracts.Services;
 using App.Contracts.ViewModels;
 using App.Core.Models;
-using App.Core.Services;
 using App.Core.Services.Interfaces;
-using App.Services;
 using CommunityToolkit.Mvvm.ComponentModel;
 using CommunityToolkit.Mvvm.Input;
 
 namespace App.ViewModels;
 
-public class DiagnoseViewModel : ObservableObject, INavigationAware
+public partial class DiagnoseViewModel : ObservableObject, INavigationAware
 {
 
     private Diagnose _selectedItem;
@@ -30,8 +27,6 @@ public class DiagnoseViewModel : ObservableObject, INavigationAware
     }
 
 
-
-
     private ObservableCollection<Diagnose> _diagnoses = new();
     public ObservableCollection<Diagnose> Diagnoses
     {
@@ -42,52 +37,44 @@ public class DiagnoseViewModel : ObservableObject, INavigationAware
     private readonly ICrudService<Diagnose> _crudService;
 
     private readonly INavigationService _navigationService;
-    public IInfoBarService InfoBarService
-    {
-        get;
-    }
 
-    public ICommand DeleteDiagnoseCommand
-    {
-        get;
-    }
+    private readonly IDialogService _dialogService;
+    public IInfoBarService InfoBarService { get; }
 
+    public ICommand DeleteDiagnoseCommand { get; }
 
-    public ICommand NavigateToUpdateDiagnoseCommand
-    {
-        get;
-    }
+    public ICommand NavigateToUpdateDiagnoseCommand { get; }
 
 
 
-    public DiagnoseViewModel(ICrudService<Diagnose> crudService, IInfoBarService infoBarService, INavigationService navigationService)
+
+    public DiagnoseViewModel(ICrudService<Diagnose> crudService, IInfoBarService infoBarService, INavigationService navigationService, IDialogService dialogService)
     {
         _crudService = crudService;
         InfoBarService = infoBarService;
         _navigationService = navigationService;
-        DeleteDiagnoseCommand = new RelayCommand(DeleteDiagnose);
+        _dialogService = dialogService;
         NavigateToUpdateDiagnoseCommand = new RelayCommand<Diagnose>(NavigateToUpdateDiagnose);
         Diagnoses = new ObservableCollection<Diagnose>();
 
     }
-
-    private async void DeleteDiagnose()
+    [RelayCommand]
+    private async void Delete()
     {
-        Diagnose diagnoseToRemove = SelectedItem;
-        Diagnoses.Remove(diagnoseToRemove);
-        await _crudService.Delete(diagnoseToRemove);
-        InfoBarService.showMessage("Erfolgreich gelöscht", "Erfolgreich");
+        var result = await _dialogService.ConfirmDeleteDialogAsync("Fehlerkategorie löschen", "Sind sie sicher, dass sie diesen Eintrag löschen wollen?");
+        if (result != null && result == true)
+        {
+            Diagnose diagnoseToRemove = SelectedItem;
+            Diagnoses.Remove(diagnoseToRemove);
+            await _crudService.Delete(diagnoseToRemove);
+            InfoBarService.showMessage("Erfolgreich gelöscht", "Erfolgreich");
+        }
 
     }
 
-    private async void NavigateToUpdateDiagnose(Diagnose diagnose)
+    private void NavigateToUpdateDiagnose(Diagnose diagnose)
     {
         _navigationService.NavigateTo("App.ViewModels.UpdateDiagnoseViewModel", diagnose);
-        /*var item = Diagnoses.FirstOrDefault(i => i == _selectedItem);
-        if (item != null)
-        {
-            item = _selectedItem;
-        }*/
 
     }
